deploy1.3.1_Transcriptions2JSON_saveDB/backend/apiTrans2JsonDicts_forDeloy1.3/deployH3_5PromptDetAndScoring/requirements.txt
pyyaml 
python-dotenv
flask
requests